requirements.txt
absl-py==1.0.0
altgraph==0.17.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
astor==0.8.1
auto-py-to-exe==2.15.0
black==21.4b2
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.25
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
docopt==0.6.2
Eel==0.12.4
fancy==3.0.0
ffprobe==0.5
fonttools==4.28.3
funcy==1.17
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.5.3
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==1.31.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.1
google-cloud-texttospeech==2.9.0
google-cloud-vision==1.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.42.0
hydra-core==1.1.1
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.2
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.3.2
layoutparser==0.3.2
Markdown==3.3.6
matplotlib==3.5.1
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
pdf2image==1.16.0
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.4.0
pipreqs==0.4.11
portalocker==2.3.2
proglog==0.1.9
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycryptodome==3.12.0
pydot==1.4.2
pydub==0.25.1
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pyparsing==3.0.6
PySimpleGUI==4.56.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sortedcontainers==2.4.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Wand==0.6.7
Werkzeug==2.0.2
whichcraft==0.6.1
yacs==0.1.8
yarg==0.1.9
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
